Make Registration an Interface and DeviceRegistration concrete class
