+++ leshan-lwm2m-server/src/main/java/org/eclipse/leshan/server/registration/Registration.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2013-2015 Sierra Wireless and others.
+ * Copyright (c) 2026 Sierra Wireless and others.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
@@ -12,130 +12,47 @@
  *
  * Contributors:
  *     Sierra Wireless - initial API and implementation
- *     Achim Kraus (Bosch Software Innovations GmbH) - use Identity as destination
- *     Michał Wadowski (Orange) - Improved compliance with rfc6690
  *******************************************************************************/
 package org.eclipse.leshan.server.registration;
 
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.Date;
 import java.util.EnumSet;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
-import java.util.TreeSet;
 
 import org.eclipse.leshan.core.LwM2m.LwM2mVersion;
 import org.eclipse.leshan.core.LwM2m.Version;
 import org.eclipse.leshan.core.endpoint.EndpointUri;
 import org.eclipse.leshan.core.link.Link;
 import org.eclipse.leshan.core.node.LwM2mPath;
-import org.eclipse.leshan.core.peer.IpPeer;
 import org.eclipse.leshan.core.peer.LwM2mPeer;
 import org.eclipse.leshan.core.request.BindingMode;
 import org.eclipse.leshan.core.request.ContentFormat;
-import org.eclipse.leshan.core.util.StringUtils;
-import org.eclipse.leshan.core.util.Validate;
 import org.eclipse.leshan.server.security.Authorizer;
 
 /**
- * An immutable structure which represent a LW-M2M client registration on the server
+ * An interface represent a LW-M2M client registration on the server
  */
-public class Registration {
+public interface Registration {
 
-    private static final long DEFAULT_LIFETIME_IN_SEC = 86400L;
+    /**
+     * @return Registration Id
+     */
+    String getId();
 
-    private final Date registrationDate;
-
-    private final LwM2mPeer clientTransportData;
-
-    private final long lifeTimeInSec;
-
-    private final String smsNumber;
-
-    private final LwM2mVersion lwM2mVersion;
-
-    private final EnumSet<BindingMode> bindingMode;
-
-    private final Boolean queueMode; // since LWM2M 1.1
-
-    // The LWM2M Client's unique end point name.
-    private final String endpoint;
-
-    private final String id;
-
-    private final Link[] objectLinks;
-
-    private final Map<String, String> additionalRegistrationAttributes;
-
-    // The location where LWM2M objects are hosted on the device
-    private final String rootPath;
-
-    // All ContentFormat supported by the client
-    private final Set<ContentFormat> supportedContentFormats;
-
-    // All supported object (object id => version)
-    private final Map<Integer, Version> supportedObjects;
-
-    // All available instances
-    private final Set<LwM2mPath> availableInstances;
-
-    private final Date lastUpdate;
-
-    private final Map<String, String> customRegistrationData;
-
-    // URI of endpoint used for this registration.
-    private final EndpointUri endpointUri;
-
-    protected Registration(Builder builder) {
-
-        // mandatory params
-        id = builder.registrationId;
-        clientTransportData = builder.clientTransportData;
-        endpoint = builder.endpoint;
-        endpointUri = builder.endpointUri;
-
-        // object links related params
-        objectLinks = builder.objectLinks;
-        rootPath = builder.rootPath;
-        supportedContentFormats = builder.supportedContentFormats;
-        supportedObjects = builder.supportedObjects;
-        availableInstances = builder.availableInstances;
-
-        // other params
-        lifeTimeInSec = builder.lifeTimeInSec;
-        lwM2mVersion = builder.lwM2mVersion;
-        bindingMode = builder.bindingMode;
-        queueMode = builder.queueMode;
-        registrationDate = builder.registrationDate;
-        lastUpdate = builder.lastUpdate;
-        smsNumber = builder.smsNumber;
-        additionalRegistrationAttributes = builder.additionalRegistrationAttributes;
-
-        customRegistrationData = builder.customRegistrationData;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public Date getRegistrationDate() {
-        return registrationDate;
-    }
+    /**
+     * @return date and time when registration was created
+     */
+    Date getRegistrationDate();
 
     /**
      * Gets the clients transport layer data.
      *
      * @return transport layer data from client's most recent registration or registration update.
      */
-    public LwM2mPeer getClientTransportData() {
-        return clientTransportData;
-    }
+    LwM2mPeer getClientTransportData();
 
     /**
      * Gets the client's network socket address.
@@ -143,12 +60,7 @@ public class Registration {
      * @return the source address from the client's most recent CoAP message. It could return {@code null} if client
      *         does not communicate over IP.
      */
-    public InetSocketAddress getSocketAddress() {
-        if (clientTransportData instanceof IpPeer) {
-            return ((IpPeer) clientTransportData).getSocketAddress();
-        }
-        return null;
-    }
+    InetSocketAddress getSocketAddress();
 
     /**
      * Gets the client's network address.
@@ -156,12 +68,7 @@ public class Registration {
      * @return the source address from the client's most recent CoAP message. It could return {@code null} if client
      *         does not communicate over IP.
      */
-    public InetAddress getAddress() {
-        if (clientTransportData instanceof IpPeer) {
-            return ((IpPeer) clientTransportData).getSocketAddress().getAddress();
-        }
-        return null;
-    }
+    InetAddress getAddress();
 
     /**
      * Gets the client's network port number.
@@ -169,140 +76,92 @@ public class Registration {
      * @return the source port from the client's most recent CoAP message. It could return {@code null} if client does
      *         not communicate over IP.
      */
-    public Integer getPort() {
-        if (clientTransportData instanceof IpPeer) {
-            return ((IpPeer) clientTransportData).getSocketAddress().getPort();
-        }
-        return null;
-    }
+    Integer getPort();
 
-    public Link[] getObjectLinks() {
-        return objectLinks;
-    }
+    /**
+     * @return array of links object of the registration
+     */
+    Link[] getObjectLinks();
 
-    public Link[] getSortedObjectLinks() {
-        // sort the list of objects
-        if (objectLinks == null) {
-            return null;
-        }
+    /**
+     * @return {@link #getObjectLinks()} but sorted by path
+     */
+    Link[] getSortedObjectLinks();
 
-        Link[] res = Arrays.copyOf(objectLinks, objectLinks.length);
+    /**
+     * @return life time in seconds of the registration
+     */
+    Long getLifeTimeInSec();
 
-        Arrays.sort(res, (o1, o2) -> {
-            if (o1 == null && o2 == null)
-                return 0;
-            if (o1 == null)
-                return -1;
-            if (o2 == null)
-                return 1;
-            // by URL
-            String[] url1 = o1.getUriReference().split("/");
-            String[] url2 = o2.getUriReference().split("/");
+    /**
+     * @return sms number linked to this registration
+     */
+    String getSmsNumber();
 
-            for (int i = 0; i < url1.length && i < url2.length; i++) {
-                // is it two numbers?
-                if (isNumber(url1[i]) && isNumber(url2[i])) {
-                    int cmp = Integer.parseInt(url1[i]) - Integer.parseInt(url2[i]);
-                    if (cmp != 0) {
-                        return cmp;
-                    }
-                } else {
+    /**
+     * @return the LWM2M version implemented by the client
+     */
+    LwM2mVersion getLwM2mVersion();
 
-                    int v = url1[i].compareTo(url2[i]);
+    /**
+     * @return the all supported binding mode
+     */
+    EnumSet<BindingMode> getBindingMode();
 
-                    if (v != 0) {
-                        return v;
-                    }
-                }
-            }
-
-            return url1.length - url2.length;
-        });
-
-        return res;
-    }
-
-    private static boolean isNumber(String s) {
-        return !StringUtils.isEmpty(s) && StringUtils.isNumeric(s);
-    }
-
-    public Long getLifeTimeInSec() {
-        return lifeTimeInSec;
-    }
-
-    public String getSmsNumber() {
-        return smsNumber;
-    }
-
-    public LwM2mVersion getLwM2mVersion() {
-        return lwM2mVersion;
-    }
-
-    public EnumSet<BindingMode> getBindingMode() {
-        return bindingMode;
-    }
-
-    public Boolean getQueueMode() {
-        return queueMode;
-    }
+    /**
+     * If you want to know if client is using queue mode regardless its LWM2M version use {@link #usesQueueMode()}
+     *
+     * @return <code>true</code> if client register with QueueMode attribute (since LWM2M 1.1)
+     */
+    Boolean getQueueMode();
 
     /**
      * @return the path where the objects are hosted on the device
      */
-    public String getRootPath() {
-        return rootPath;
-    }
+    String getRootPath();
 
     /**
      * @return all {@link ContentFormat} supported by the client.
      */
-    public Set<ContentFormat> getSupportedContentFormats() {
-        return supportedContentFormats;
-    }
+    Set<ContentFormat> getSupportedContentFormats();
 
     /**
      * @return all available object instance by the client
      */
-    public Set<LwM2mPath> getAvailableInstances() {
-        return availableInstances;
-    }
+    Set<LwM2mPath> getAvailableInstances();
 
     /**
      * Gets the unique name the client has registered with.
      *
      * @return the name
      */
-    public String getEndpoint() {
-        return endpoint;
-    }
+    String getEndpoint();
 
-    public Date getLastUpdate() {
-        return lastUpdate;
-    }
+    /**
+     * @return date/time of the last registration update
+     */
+    Date getLastUpdate();
 
-    public long getExpirationTimeStamp() {
-        return getExpirationTimeStamp(0L);
-    }
+    /**
+     * @return timestamp when the registration will expire (in Milliseconds).
+     */
+    long getExpirationTimeStamp();
 
-    public long getExpirationTimeStamp(long gracePeriodInSec) {
-        return lastUpdate.getTime() + lifeTimeInSec * 1000 + gracePeriodInSec * 1000;
-    }
+    /**
+     * @return timestamp when the registration will expire (in Milliseconds) but with a grace period in seconds allowing
+     *         some registration update lateness.
+     */
+    long getExpirationTimeStamp(long gracePeriodInSec);
 
     /**
      * @return True if DTLS handshake can be initiated by the Server for this registration.
      */
-    public boolean canInitiateConnection() {
-        // We consider that initiates a connection (acting as DTLS client to initiate a handshake) does not make sense
-        // for QueueMode as if we lost connection device is probably absent.
-        return !usesQueueMode();
-    }
+    boolean canInitiateConnection();
 
     /**
      * @return true if the last registration update was done less than lifetime seconds ago.
      */
-    public boolean isAlive() {
-        return isAlive(0);
-    }
+    boolean isAlive();
 
     /**
      * This is the same idea than {@link Registration#isAlive()} but with a grace period. <br>
@@ -310,273 +169,37 @@ public class Registration {
      * @param gracePeriodInSec an extra time for the registration lifetime.
      * @return true if the last registration update was done less than lifetime+gracePeriod seconds ago.
      */
-    public boolean isAlive(long gracePeriodInSec) {
-        return getExpirationTimeStamp(gracePeriodInSec) > System.currentTimeMillis();
-    }
+    boolean isAlive(long gracePeriodInSec);
 
-    public Map<String, String> getAdditionalRegistrationAttributes() {
-        return additionalRegistrationAttributes;
-    }
+    /**
+     * @return some additional attribute present at registration time but not part of the LWM2M specification.
+     */
+    Map<String, String> getAdditionalRegistrationAttributes();
 
-    public boolean usesQueueMode() {
-        if (lwM2mVersion.olderThan(LwM2mVersion.V1_1))
-            return bindingMode.contains(BindingMode.Q);
-        else
-            return queueMode;
-    }
+    /**
+     * @return <code>true</code> if client is using Queue mode.
+     */
+    boolean usesQueueMode();
 
     /**
      * @param objectid the object id for which we want to know the supported version.
      * @return the supported version of the object with the id {@code objectid}. If the object is not supported return
      *         {@code null}
      */
-    public Version getSupportedVersion(Integer objectid) {
-        return getSupportedObject().get(objectid);
-    }
+    Version getSupportedVersion(Integer objectid);
 
     /**
      * @return a map from {@code objectId} {@literal =>} {@code supportedVersion} for each supported objects. supported.
      */
-    public Map<Integer, Version> getSupportedObject() {
-        return supportedObjects;
-    }
+    Map<Integer, Version> getSupportedObject();
 
     /**
      * @return Some custom registration data which could have been added at Registration by the {@link Authorizer}
      */
-    public Map<String, String> getCustomRegistrationData() {
-        return customRegistrationData;
-    }
+    Map<String, String> getCustomRegistrationData();
 
     /**
      * @return URI of the server endpoint used by client to register.
      */
-    public EndpointUri getEndpointUri() {
-        return endpointUri;
-    }
-
-    @Override
-    public String toString() {
-        return String.format(
-                "Registration [registrationDate=%s, clientTransportData=%s, lifeTimeInSec=%s, smsNumber=%s, lwM2mVersion=%s, bindingMode=%s, queueMode=%s, endpoint=%s, id=%s, objectLinks=%s, additionalRegistrationAttributes=%s, rootPath=%s, supportedContentFormats=%s, supportedObjects=%s, availableInstances=%s, lastUpdate=%s, customRegistrationData=%s, endpointUri=%s]",
-                registrationDate, clientTransportData, lifeTimeInSec, smsNumber, lwM2mVersion, bindingMode, queueMode,
-                endpoint, id, Arrays.toString(objectLinks), additionalRegistrationAttributes, rootPath,
-                supportedContentFormats, supportedObjects, availableInstances, lastUpdate, customRegistrationData,
-                endpointUri);
-    }
-
-    @Override
-    public final boolean equals(Object o) {
-        if (this == o)
-            return true;
-        if (!(o instanceof Registration))
-            return false;
-        Registration that = (Registration) o;
-        return lifeTimeInSec == that.lifeTimeInSec && Objects.equals(registrationDate, that.registrationDate)
-                && Objects.equals(clientTransportData, that.clientTransportData)
-                && Objects.equals(smsNumber, that.smsNumber) && Objects.equals(lwM2mVersion, that.lwM2mVersion)
-                && Objects.equals(bindingMode, that.bindingMode) && Objects.equals(queueMode, that.queueMode)
-                && Objects.equals(endpoint, that.endpoint) && Objects.equals(id, that.id)
-                && Arrays.equals(objectLinks, that.objectLinks)
-                && Objects.equals(additionalRegistrationAttributes, that.additionalRegistrationAttributes)
-                && Objects.equals(rootPath, that.rootPath)
-                && Objects.equals(supportedContentFormats, that.supportedContentFormats)
-                && Objects.equals(supportedObjects, that.supportedObjects)
-                && Objects.equals(availableInstances, that.availableInstances)
-                && Objects.equals(lastUpdate, that.lastUpdate)
-                && Objects.equals(customRegistrationData, that.customRegistrationData)
-                && Objects.equals(endpointUri, that.endpointUri);
-    }
-
-    @Override
-    public final int hashCode() {
-        return Objects.hash(registrationDate, clientTransportData, lifeTimeInSec, smsNumber, lwM2mVersion, bindingMode,
-                queueMode, endpoint, id, Arrays.hashCode(objectLinks), additionalRegistrationAttributes, rootPath,
-                supportedContentFormats, supportedObjects, availableInstances, lastUpdate, customRegistrationData,
-                endpointUri);
-    }
-
-    public static class Builder {
-        private final String registrationId;
-        private final String endpoint;
-        private final LwM2mPeer clientTransportData;
-        private final EndpointUri endpointUri;
-
-        private Date registrationDate;
-        private Date lastUpdate;
-        private Long lifeTimeInSec;
-        private String smsNumber;
-        private EnumSet<BindingMode> bindingMode;
-        private Boolean queueMode;
-        private LwM2mVersion lwM2mVersion = LwM2mVersion.getDefault();
-        private Link[] objectLinks;
-        private String rootPath;
-        private Set<ContentFormat> supportedContentFormats;
-        private Map<Integer, Version> supportedObjects;
-        private Set<LwM2mPath> availableInstances;
-        private Map<String, String> additionalRegistrationAttributes;
-        private Map<String, String> customRegistrationData;
-
-        public Builder(Registration registration) {
-
-            // mandatory params
-            registrationId = registration.id;
-            clientTransportData = registration.clientTransportData;
-            endpoint = registration.endpoint;
-            endpointUri = registration.endpointUri;
-
-            // object links related params
-            objectLinks = registration.objectLinks;
-            rootPath = registration.rootPath;
-            supportedContentFormats = registration.supportedContentFormats;
-            supportedObjects = registration.supportedObjects;
-            availableInstances = registration.availableInstances;
-
-            // other params
-            lifeTimeInSec = registration.lifeTimeInSec;
-            lwM2mVersion = registration.lwM2mVersion;
-            bindingMode = registration.bindingMode;
-            queueMode = registration.queueMode;
-            registrationDate = registration.registrationDate;
-            lastUpdate = registration.lastUpdate;
-            smsNumber = registration.smsNumber;
-            additionalRegistrationAttributes = registration.additionalRegistrationAttributes;
-
-            customRegistrationData = registration.customRegistrationData;
-        }
-
-        public Builder(String registrationId, String endpoint, LwM2mPeer clientTransportData, EndpointUri endpointUri) {
-
-            Validate.notNull(registrationId);
-            Validate.notEmpty(endpoint);
-            Validate.notNull(clientTransportData);
-            Validate.notNull(endpointUri);
-
-            this.registrationId = registrationId;
-            this.endpoint = endpoint;
-            this.clientTransportData = clientTransportData;
-            this.endpointUri = endpointUri;
-        }
-
-        public Builder registrationDate(Date registrationDate) {
-            this.registrationDate = registrationDate;
-            return this;
-        }
-
-        public Builder lastUpdate(Date lastUpdate) {
-            this.lastUpdate = lastUpdate;
-            return this;
-        }
-
-        public Builder lifeTimeInSec(Long lifetimeInSec) {
-            this.lifeTimeInSec = lifetimeInSec;
-            return this;
-        }
-
-        public Builder smsNumber(String smsNumber) {
-            this.smsNumber = smsNumber;
-            return this;
-        }
-
-        public Builder bindingMode(EnumSet<BindingMode> bindingMode) {
-            this.bindingMode = bindingMode;
-            return this;
-        }
-
-        public Builder queueMode(Boolean queueMode) {
-            this.queueMode = queueMode;
-            return this;
-        }
-
-        public Builder lwM2mVersion(LwM2mVersion lwM2mVersion) {
-            this.lwM2mVersion = lwM2mVersion;
-            return this;
-        }
-
-        public Builder objectLinks(Link[] objectLinks) {
-            this.objectLinks = objectLinks;
-            return this;
-        }
-
-        public Builder rootPath(String rootPath) {
-            this.rootPath = rootPath;
-            return this;
-        }
-
-        public Builder supportedContentFormats(Set<ContentFormat> supportedContentFormats) {
-            this.supportedContentFormats = supportedContentFormats;
-            return this;
-        }
-
-        public Builder supportedContentFormats(ContentFormat... supportedContentFormats) {
-            this.supportedContentFormats = new HashSet<>();
-            for (ContentFormat contentFormat : supportedContentFormats) {
-                this.supportedContentFormats.add(contentFormat);
-            }
-            return this;
-        }
-
-        public Builder supportedObjects(Map<Integer, Version> supportedObjects) {
-            this.supportedObjects = supportedObjects;
-            return this;
-        }
-
-        public Builder availableInstances(Set<LwM2mPath> availableInstances) {
-            this.availableInstances = availableInstances;
-            return this;
-        }
-
-        public Builder additionalRegistrationAttributes(Map<String, String> additionalRegistrationAttributes) {
-            this.additionalRegistrationAttributes = additionalRegistrationAttributes;
-            return this;
-        }
-
-        public Builder customRegistrationData(Map<String, String> customRegistrationData) {
-            this.customRegistrationData = customRegistrationData;
-            return this;
-        }
-
-        public Registration build() {
-            // Define Default value
-            rootPath = rootPath == null ? "/" : rootPath;
-            lifeTimeInSec = lifeTimeInSec == null ? DEFAULT_LIFETIME_IN_SEC : lifeTimeInSec;
-            lwM2mVersion = lwM2mVersion == null ? LwM2mVersion.getDefault() : lwM2mVersion;
-            bindingMode = bindingMode == null ? EnumSet.of(BindingMode.U) : bindingMode;
-            queueMode = queueMode == null && lwM2mVersion.newerThan(LwM2mVersion.V1_0) ? Boolean.FALSE : queueMode;
-            registrationDate = registrationDate == null ? new Date() : registrationDate;
-            lastUpdate = lastUpdate == null ? new Date() : lastUpdate;
-
-            // Make collection immutable
-            // We create a new Collection and make it "unmodifiable".
-            if (supportedContentFormats == null || supportedContentFormats.isEmpty()) {
-                supportedContentFormats = Collections.emptySet();
-            } else {
-                supportedContentFormats = Collections.unmodifiableSet(new HashSet<>(supportedContentFormats));
-            }
-            if (supportedObjects == null || supportedObjects.isEmpty()) {
-                supportedObjects = Collections.emptyMap();
-            } else {
-                supportedObjects = Collections.unmodifiableMap(new HashMap<>(supportedObjects));
-            }
-            if (availableInstances == null || availableInstances.isEmpty()) {
-                availableInstances = Collections.emptySet();
-            } else {
-                availableInstances = Collections.unmodifiableSet(new TreeSet<>(availableInstances));
-            }
-            if (additionalRegistrationAttributes == null || additionalRegistrationAttributes.isEmpty()) {
-                additionalRegistrationAttributes = Collections.emptyMap();
-            } else {
-                additionalRegistrationAttributes = Collections
-                        .unmodifiableMap(new HashMap<>(additionalRegistrationAttributes));
-            }
-            if (customRegistrationData == null || customRegistrationData.isEmpty()) {
-                customRegistrationData = Collections.emptyMap();
-            } else {
-                customRegistrationData = Collections.unmodifiableMap(new HashMap<>(customRegistrationData));
-            }
-
-            // Create Registration
-            return new Registration(this);
-        }
-    }
+    EndpointUri getEndpointUri();
 }
