+++ leshan-lwm2m-server/src/main/java/org/eclipse/leshan/server/registration/RegistrationHandler.java
@@ -82,7 +82,7 @@ public class RegistrationHandler {
         }
 
         // Create Registration from RegisterRequest
-        Registration.Builder builder = new Registration.Builder(
+        DeviceRegistration.Builder builder = new DeviceRegistration.Builder(
                 registrationIdProvider.getRegistrationId(registerRequest), endpointName, sender, endpointUsed);
 
         builder.lwM2mVersion(lwM2mVersion) //
@@ -98,7 +98,7 @@ public class RegistrationHandler {
                 .supportedObjects(objLinksData.getSupportedObjects()) //
                 .availableInstances(objLinksData.getAvailableInstances());
 
-        Registration registrationToApproved = builder.build();
+        DeviceRegistration registrationToApproved = builder.build();
 
         // We check if the client get authorization.
         Authorization authorization = authorizer.isAuthorized(registerRequest, registrationToApproved, sender,
@@ -110,7 +110,7 @@ public class RegistrationHandler {
         // Add Authorization Custom Data to Registration if needed
         final Registration approvedRegistration;
         if (authorization.hasCustomData()) {
-            approvedRegistration = new Registration.Builder(registrationToApproved)
+            approvedRegistration = new DeviceRegistration.Builder(registrationToApproved)
                     .customRegistrationData(authorization.getCustomData()).build();
         } else {
             approvedRegistration = registrationToApproved;
