+++ leshan-lwm2m-server/src/test/java/org/eclipse/leshan/server/queue/PresenceServiceTest.java
@@ -27,6 +27,7 @@ import java.util.EnumSet;
 
 import org.eclipse.leshan.core.peer.IpPeer;
 import org.eclipse.leshan.core.request.BindingMode;
+import org.eclipse.leshan.server.registration.DeviceRegistration;
 import org.eclipse.leshan.server.registration.Registration;
 import org.junit.jupiter.api.Test;
 
@@ -67,7 +68,7 @@ public class PresenceServiceTest {
     }
 
     private Registration givenASimpleClient() throws UnknownHostException {
-        Registration.Builder builder = new Registration.Builder("ID", "urn:client",
+        DeviceRegistration.Builder builder = new DeviceRegistration.Builder("ID", "urn:client",
                 new IpPeer(new InetSocketAddress(Inet4Address.getLoopbackAddress(), 12354)),
                 uriHandler.createUri("coap://localhost:5683"));
 
@@ -78,7 +79,7 @@ public class PresenceServiceTest {
 
     private Registration givenASimpleClientWithQueueMode() throws UnknownHostException {
 
-        Registration.Builder builder = new Registration.Builder("ID", "urn:client",
+        DeviceRegistration.Builder builder = new DeviceRegistration.Builder("ID", "urn:client",
                 new IpPeer(new InetSocketAddress(Inet4Address.getLoopbackAddress(), 12354)),
                 uriHandler.createUri("coap://localhost:5683"));
