+++ leshan-lwm2m-server/src/test/java/org/eclipse/leshan/server/registration/RegistrationTest.java
@@ -36,7 +36,7 @@ import org.eclipse.leshan.core.link.lwm2m.DefaultLwM2mLinkParser;
 import org.eclipse.leshan.core.node.LwM2mPath;
 import org.eclipse.leshan.core.peer.IpPeer;
 import org.eclipse.leshan.core.request.ContentFormat;
-import org.eclipse.leshan.server.registration.Registration.Builder;
+import org.eclipse.leshan.server.registration.DeviceRegistration.Builder;
 import org.eclipse.leshan.server.registration.RegistrationDataExtractor.RegistrationData;
 import org.junit.jupiter.api.Test;
 
@@ -279,7 +279,7 @@ public class RegistrationTest {
 
     private Registration given_a_registration_with_object_link_like(String objectLinks, LwM2mVersion version)
             throws LinkParseException {
-        Builder builder = new Registration.Builder("id", "endpoint",
+        Builder builder = new DeviceRegistration.Builder("id", "endpoint",
                 new IpPeer(InetSocketAddress.createUnresolved("localhost", 0)),
                 uriHandler.createUri("coap://localhost:5683"));
