+++ leshan-tl-cf-server-coap/src/test/java/org/eclipse/leshan/transport/californium/server/request/CoapRequestBuilderTest.java
@@ -56,8 +56,9 @@ import org.eclipse.leshan.core.tlv.Tlv;
 import org.eclipse.leshan.core.tlv.Tlv.TlvType;
 import org.eclipse.leshan.core.tlv.TlvDecoder;
 import org.eclipse.leshan.server.registration.DefaultRegistrationDataExtractor;
+import org.eclipse.leshan.server.registration.DeviceRegistration;
+import org.eclipse.leshan.server.registration.DeviceRegistration.Builder;
 import org.eclipse.leshan.server.registration.Registration;
-import org.eclipse.leshan.server.registration.Registration.Builder;
 import org.eclipse.leshan.server.registration.RegistrationDataExtractor.RegistrationData;
 import org.eclipse.leshan.transport.californium.identity.DefaultCoapIdentityHandler;
 import org.eclipse.leshan.transport.californium.identity.IdentityHandler;
@@ -85,7 +86,7 @@ public class CoapRequestBuilderTest {
     }
 
     private Registration newRegistration(String rootpath) throws UnknownHostException {
-        Builder b = new Registration.Builder("regid", "endpoint",
+        Builder b = new DeviceRegistration.Builder("regid", "endpoint",
                 new IpPeer(new InetSocketAddress(Inet4Address.getLoopbackAddress(), 12354)),
                 uriHandler.createUri("coap://localhost:5683"));
